# Remove allocations in ByteKey.  Pre-allocate size for connections slice in mergeConnections