@DataDog/burrito 

Saw `mergeConnections` as a large source of allocations.  Pre-allocating the list should help reduce that.

I also removed some low hanging fruit in `ByteKey`

```
BenchmarkOldByteKey-2   	20000000	        81.0 ns/op	      24 B/op	       3 allocs/op
BenchmarkNewByteKey-2   	50000000	        24.8 ns/op	       0 B/op	       0 allocs/op
```