@sunhay 2019-04-15
We do have some benchmarks in `state.go`, can you run them and provide output?